generate, validate: isolate gojson* dependencies.
Split out those few capability validation functions
(LastCap(), CapValid()) which Generator depends on
into a validate/capabilities subpackage of their own.
This should prevent github.com/xeipuuv/gojson* from
sneaking in to the dependencies of anyone who uses
Generator for OCI Spec manipulation.

Those gojsonschema and related packages are often
considered problematic because they do not seem to
be actively maintained. At the time of this commit
they received their last updates in 2018, 2019 and
2020.
